separated spec files for icon
